Fix a integer trauction issue - calculating the current time in nanoseconds in 32-bit expression would cause pthread_cond_timedwait to time out immediately.  Add explicit casts to the TimeValue::TimeValue ctor that takes a struct timeval and change the NanoSecsPerSec etc constants defined in TimeValue to be uint64_t so any other calculations involving these should be promoted to 64-bit even when lldb is built for 32-bit.
<rdar://problem/11204073>, <rdar://problem/11179821>, <rdar://problem/11194705>.

llvm-svn: 154250

+++ lldb/include/lldb/Host/TimeValue.h
@@ -30,9 +30,9 @@ namespace lldb_private {
 class TimeValue
 {
 public:
-    static const uint32_t MicroSecPerSec = 1000000UL;
-    static const uint32_t NanoSecPerSec = 1000000000UL;
-    static const uint32_t NanoSecPerMicroSec = 1000U;
+    static const uint64_t MicroSecPerSec = 1000000UL;
+    static const uint64_t NanoSecPerSec = 1000000000UL;
+    static const uint64_t NanoSecPerMicroSec = 1000U;
 
     //------------------------------------------------------------------
     // Constructors and Destructors
